move to common pkg, surface listening errors, use context
