+++ cmd/agent/app/start.go
@@ -6,6 +6,7 @@
 package app
 
 import (
+	"context"
 	"fmt"
 	"runtime"
 	"strings"
@@ -20,10 +21,10 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/api"
-	"github.com/DataDog/datadog-agent/cmd/agent/api/healthport"
 	"github.com/DataDog/datadog-agent/cmd/agent/common"
 	"github.com/DataDog/datadog-agent/cmd/agent/gui"
 	"github.com/DataDog/datadog-agent/pkg/aggregator"
+	"github.com/DataDog/datadog-agent/pkg/api/healthport"
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/embed/jmx"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/dogstatsd"
@@ -74,6 +75,8 @@ func start(cmd *cobra.Command, args []string) error {
 
 // StartAgent Initializes the agent process
 func StartAgent() error {
+	// Main context passed to components
+	common.MainCtx, common.MainCtxCancel = context.WithCancel(context.Background())
 
 	// Global Agent configuration
 	err := common.SetupConfig(confFilePath)
@@ -142,12 +145,11 @@ func StartAgent() error {
 	// Setup healthcheck port
 	var healthPort = config.Datadog.GetInt("health_port")
 	if healthPort > 0 {
-		var err error
-		common.HealthServer, err = healthport.NewHealthServer(healthPort)
+		err := healthport.Serve(common.MainCtx, healthPort)
 		if err != nil {
 			return log.Errorf("Error starting health port, exiting: %v", err)
 		}
-		go common.HealthServer.ListenAndServe()
+		log.Debugf("Health check listening on port %d", healthPort)
 	}
 
 	if pidfilePath != "" {
@@ -305,6 +307,8 @@ func StopAgent() {
 	}
 
 	// gracefully shut down any component
+	common.MainCtxCancel()
+
 	if common.DSD != nil {
 		common.DSD.Stop()
 	}
@@ -318,9 +322,6 @@ func StopAgent() {
 	jmx.StopJmxfetch()
 	if common.Forwarder != nil {
 		common.Forwarder.Stop()
-	}
-	if common.HealthServer != nil {
-		common.HealthServer.Close()
 	}
 	logs.Stop()
 	gui.StopGUIServer()
